refactor(ui): consolidate the three formatCurrency copies into one util
budgetModal.ts, toolInfo.ts, and budgetInsightsRule.ts each carried an
independent formatCurrency implementation, copy-pasted forward over
several months — which is how the same $10-$99 decimal bug (#146)
ended up duplicated in all three. Extracts the canonical version to
app/src/utils/currency.ts and points all three call sites at it.

toolInfo.ts's copy also rounded the signed value instead of its
absolute value in the >=$100 branch, and its two `?? '$0'` fallbacks
were dead code (cost/maintenance are never undefined at those call
sites) — both cleaned up as part of the consolidation.

Replaces the three per-file formatCurrency test files with one
consolidated suite at app/src/utils/currency.test.ts.
